Rework MNIST loader.
The loader now splits the data IID as default but can perform a LDA with specified alpha parameter. Further to ensure a Uniform split we now trim away data from the test and train sets such that the data contain the same amount of labels AND such that it can be split into the desired number of clients.  As such the number of clients now can make the total amount of data to vary.