Make client momentum possible
With this implementation a momentum for the client optimizer is possible during each round. The momentum is thus reset between each round. This seems reasonable as we generally sample a subset of clients for each round, as such there is no guarantee that the next time a client appear the momentums calculated the last time are valid. Further keeping one optimized for each client is too memory consuming and sharing optimizer between clients does not work as the data is non-iid thus the momentum would not be the same between clients.