Add plotting notebook example to easily illustrate results
The notebook assume you to change the path to the logs manually. It also introduces a new package that aim to strip the metadata from notebooks so that hosting notebooks on git is reasonable. Otherwise for example images will be embeded as SVG and change between every commit --> huge diffs.