Implement Stackoverflow model
Update the model and dataloader so that the dimensions match and it is possible to run. Also modify/hack the callbacks such that when we do NWP we calculate the accuracy etc. on only the last token presented. This is done by an if statement simply removing the other tokens when the callback notice that the target has too many dimensions.